gtable-package {gtable}	R Documentation

gtable: Arrange 'Grobs' in Tables

Description

Tools to make it easier to work with "tables" of 'grobs'. The 'gtable' package defines a 'gtable' grob class that specifies a grid along with a list of grobs and their placement in the grid. Further the package makes it easy to manipulate and combine 'gtable' objects so that complex compositions can be built up sequentially.

Author(s)

Maintainer: Thomas Lin Pedersen thomas.pedersen@posit.co

Authors:

Hadley Wickham hadley@posit.co

Other contributors:

Posit Software, PBC [copyright holder]

See Also

Useful links:

https://gtable.r-lib.org

https://github.com/r-lib/gtable

Report bugs at https://github.com/r-lib/gtable/issues

[Package gtable version 0.3.4 Index]