gtable_col {gtable}	R Documentation

Create a single column gtable.

Description

Create a single column gtable.

Usage

  gtable_col(name, grobs, width = NULL, heights = NULL,
    z = NULL, vp = NULL)

Arguments

width
a unit vector giving the width of this column

vp
a grid viewport object (or NULL).

name
a string giving the name of the table. This is used to name the layout viewport

heights
a unit vector giving the height of each row

grobs
a single grob or a list of grobs

z
a numeric vector giving the order in which the grobs should be plotted. Use Inf (the default) to plot above or -Inf below all existing grobs. By default positions are on the integers, giving plenty of room to insert new grobs between existing grobs.

Examples

a <- rectGrob(gp = gpar(fill = "red"))
b <- circleGrob()
c <- linesGrob()
gt <- gtable_col("demo", list(a, b, c))
gt
plot(gt)
gtable_show_layout(gt)

[Package gtable version 0.1.2 Index]