gtable_row {gtable}	R Documentation

Create a single row gtable.

Description

This function puts grobs in a list side-by-side in a single-row gtable from left to right witrh the given widths and height.

Usage

gtable_row(name, grobs, height = NULL, widths = NULL, z = NULL, vp = NULL)

Arguments

name
a string giving the name of the table. This is used to name the layout viewport

grobs
a single grob or a list of grobs

height
a unit vector giving the height of this row

widths
a unit vector giving the width of each column

z
a numeric vector giving the order in which the grobs should be plotted. Use Inf (the default) to plot above or -Inf below all existing grobs. By default positions are on the integers, giving plenty of room to insert new grobs between existing grobs.

vp
a grid viewport object (or NULL).

Value

A gtable with a single row and the same number of columns as elements in the grobs list

See Also

Other gtable construction: gtable_col(), gtable_matrix(), gtable_spacer, gtable()

Examples

library(grid)
a <- rectGrob(gp = gpar(fill = "red"))
b <- circleGrob()
c <- linesGrob()
gt <- gtable_row("demo", list(a, b, c))
gt
plot(gt)
gtable_show_layout(gt)

[Package gtable version 0.3.4 Index]