z_normalise {gtable}	R Documentation

Normalise z values within a gtable object

Description

The z values within a gtable object can be any numeric values. This function will change them to integers (starting from 1), preserving the original order.

Usage

z_normalise(x, i = 1)

Arguments

x
A gtable object

i
The z value to start counting up from (default is 1)

Details

Ties are handled by the "first" method: the first occurrence of a value wins.

[Package gtable version 0.2.0 Index]