bind {gtable}	R Documentation

Row and column binding for gtables.

Description

Row and column binding for gtables.

Usage

## S3 method for class 'gtable'
rbind(..., size = "max", z = NULL)

## S3 method for class 'gtable'
cbind(..., size = "max", z = NULL)

Arguments

...
gtables to combine (x and y)

size
How should the widths (for rbind) and the heights (for cbind) be combined across the gtables: take values from first, or last gtable, or compute the min or max values. Defaults to max.

z
A numeric vector indicating the relative z values of each gtable. The z values of each object in the resulting gtable will be modified to fit this order. If NULL, then the z values of obects within each gtable will not be modified.

[Package gtable version 0.2.0 Index]